evonote/requirements.txt
numpy==1.25.0
plotly==5.9.0
PyHyphen==4.0.3
